+++ src/oci/core/models/export_image_via_object_storage_uri_details.py
@@ -14,6 +14,10 @@ class ExportImageViaObjectStorageUriDetails(ExportImageDetails):
         Initializes a new ExportImageViaObjectStorageUriDetails object with values from values from keyword arguments. The
         following keyword arguments are supported (corresponding to the getters/setters of this class):
 
+        :param destination_type:
+            The value to assign to the destination_type property of this ExportImageViaObjectStorageUriDetails.
+        :type destination_type: str
+
         :param destination_uri:
             The value to assign to the destination_uri property of this ExportImageViaObjectStorageUriDetails.
         :type destination_uri: str