+++ src/oci/core/models/image_source_via_object_storage_uri_details.py
@@ -14,6 +14,10 @@ class ImageSourceViaObjectStorageUriDetails(ImageSourceDetails):
         Initializes a new ImageSourceViaObjectStorageUriDetails object with values from values from keyword arguments. The
         following keyword arguments are supported (corresponding to the getters/setters of this class):
 
+        :param source_type:
+            The value to assign to the source_type property of this ImageSourceViaObjectStorageUriDetails.
+        :type source_type: str
+
         :param source_uri:
             The value to assign to the source_uri property of this ImageSourceViaObjectStorageUriDetails.
         :type source_uri: str