+++ src/oci/core/models/launch_instance_details.py
@@ -54,6 +54,10 @@ class LaunchInstanceDetails(object):
             The value to assign to the shape property of this LaunchInstanceDetails.
         :type shape: str
 
+        :param source_details:
+            The value to assign to the source_details property of this LaunchInstanceDetails.
+        :type source_details: InstanceSourceDetails
+
         :param subnet_id:
             The value to assign to the subnet_id property of this LaunchInstanceDetails.
         :type subnet_id: str
@@ -70,6 +74,7 @@ class LaunchInstanceDetails(object):
             'ipxe_script': 'str',
             'metadata': 'dict(str, str)',
             'shape': 'str',
+            'source_details': 'InstanceSourceDetails',
             'subnet_id': 'str'
         }
 
@@ -84,6 +89,7 @@ class LaunchInstanceDetails(object):
             'ipxe_script': 'ipxeScript',
             'metadata': 'metadata',
             'shape': 'shape',
+            'source_details': 'sourceDetails',
             'subnet_id': 'subnetId'
         }
 
@@ -97,6 +103,7 @@ class LaunchInstanceDetails(object):
         self._ipxe_script = None
         self._metadata = None
         self._shape = None
+        self._source_details = None
         self._subnet_id = None
 
     @property
@@ -243,7 +250,7 @@ class LaunchInstanceDetails(object):
     def hostname_label(self):
         """
         Gets the hostname_label of this LaunchInstanceDetails.
-        Deprecated. Instead use `hostnameLabel` in
+        Deprecated. Instead Use `hostnameLabel` in
         :class:`CreateVnicDetails`.
         If you provide both, the values must match.
 
@@ -257,7 +264,7 @@ class LaunchInstanceDetails(object):
     def hostname_label(self, hostname_label):
         """
         Sets the hostname_label of this LaunchInstanceDetails.
-        Deprecated. Instead use `hostnameLabel` in
+        Deprecated. Instead Use `hostnameLabel` in
         :class:`CreateVnicDetails`.
         If you provide both, the values must match.
 
@@ -271,7 +278,8 @@ class LaunchInstanceDetails(object):
     def image_id(self):
         """
         Gets the image_id of this LaunchInstanceDetails.
-        The OCID of the image used to boot the instance.
+        Deprecated. Use `sourceDetails` with :func:`instance_source_via_image_details`
+        source type instead. If you specify values for both, the values must match.
 
 
         :return: The image_id of this LaunchInstanceDetails.
@@ -283,7 +291,8 @@ class LaunchInstanceDetails(object):
     def image_id(self, image_id):
         """
         Sets the image_id of this LaunchInstanceDetails.
-        The OCID of the image used to boot the instance.
+        Deprecated. Use `sourceDetails` with :func:`instance_source_via_image_details`
+        source type instead. If you specify values for both, the values must match.
 
 
         :param image_id: The image_id of this LaunchInstanceDetails.
@@ -538,6 +547,30 @@ class LaunchInstanceDetails(object):
         :type: str
         """
         self._shape = shape
+
+    @property
+    def source_details(self):
+        """
+        Gets the source_details of this LaunchInstanceDetails.
+        Details for creating an instance.
+
+
+        :return: The source_details of this LaunchInstanceDetails.
+        :rtype: InstanceSourceDetails
+        """
+        return self._source_details
+
+    @source_details.setter
+    def source_details(self, source_details):
+        """
+        Sets the source_details of this LaunchInstanceDetails.
+        Details for creating an instance.
+
+
+        :param source_details: The source_details of this LaunchInstanceDetails.
+        :type: InstanceSourceDetails
+        """
+        self._source_details = source_details
 
     @property
     def subnet_id(self):