+++ src/oci/core/models/volume_source_from_volume_details.py
@@ -14,6 +14,10 @@ class VolumeSourceFromVolumeDetails(VolumeSourceDetails):
         Initializes a new VolumeSourceFromVolumeDetails object with values from values from keyword arguments. The
         following keyword arguments are supported (corresponding to the getters/setters of this class):
 
+        :param type:
+            The value to assign to the type property of this VolumeSourceFromVolumeDetails.
+        :type type: str
+
         :param id:
             The value to assign to the id property of this VolumeSourceFromVolumeDetails.
         :type id: str