+++ src/oci/load_balancer/models/work_request.py
@@ -132,6 +132,8 @@ class WorkRequest(object):
     def lifecycle_state(self):
         """
         Gets the lifecycle_state of this WorkRequest.
+        The current state of the work request.
+
         Allowed values for this property are: "ACCEPTED", "IN_PROGRESS", "FAILED", "SUCCEEDED", 'UNKNOWN_ENUM_VALUE'.
         Any unrecognized values returned by a service will be mapped to 'UNKNOWN_ENUM_VALUE'.
 
@@ -145,6 +147,8 @@ class WorkRequest(object):
     def lifecycle_state(self, lifecycle_state):
         """
         Sets the lifecycle_state of this WorkRequest.
+        The current state of the work request.
+
 
         :param lifecycle_state: The lifecycle_state of this WorkRequest.
         :type: str